+++ .eslintrc.js
@@ -44,5 +44,7 @@ module.exports = {
   ],
   rules: {
     '@sentry-internal/sdk/no-async-await': 'off',
+    '@sentry-internal/sdk/no-optional-chaining': 'off',
+    '@sentry-internal/sdk/no-nullish-coalescing': 'off',
   },
 };